Relocate naive gate run
