+++ src/missions/gate.rs
@@ -13,14 +13,52 @@ use crate::{
 use super::{
     action::{
         wrap_action, ActionChain, ActionConcurrent, ActionExec, ActionMod, ActionSequence,
-        ActionWhile, FirstValid,
+        ActionWhile, FirstValid, TupleSecond,
     },
     action_context::{GetControlBoard, GetFrontCamMat, GetMainElectronicsBoard},
     basic::descend_and_go_forward,
     comms::StartBno055,
-    movement::{CountFalse, CountTrue},
-    vision::{DetectTarget, VisionNorm},
+    movement::{AdjustMovementAngle, CountFalse, CountTrue},
+    vision::{DetectTarget, VisionNorm, VisionNormOffset},
 };
+
+pub fn gate_run_naive<
+    Con: Send
+        + Sync
+        + GetControlBoard<WriteHalf<SerialStream>>
+        + GetMainElectronicsBoard
+        + GetFrontCamMat,
+>(
+    context: &Con,
+) -> impl ActionExec<()> + '_ {
+    let depth: f32 = -1.0;
+
+    ActionSequence::new(
+        ActionConcurrent::new(descend_and_go_forward(context), StartBno055::new(context)),
+        ActionSequence::new(
+            ActionWhile::new(ActionChain::new(
+                VisionNormOffset::<Con, GatePoles<OnnxModel>, f64>::new(
+                    context,
+                    GatePoles::default(),
+                ),
+                TupleSecond::new(ActionConcurrent::new(
+                    AdjustMovementAngle::new(context, depth),
+                    CountTrue::new(3),
+                )),
+            )),
+            ActionWhile::new(ActionChain::new(
+                VisionNormOffset::<Con, GatePoles<OnnxModel>, f64>::new(
+                    context,
+                    GatePoles::default(),
+                ),
+                TupleSecond::new(ActionConcurrent::new(
+                    AdjustMovementAngle::new(context, depth),
+                    CountFalse::new(10),
+                )),
+            )),
+        ),
+    )
+}
 
 pub fn gate_run_complex<
     Con: Send
