+++ src/missions/mod.rs
@@ -5,6 +5,7 @@ pub mod buoy_circle;
 pub mod buoy_hit;
 pub mod comms;
 pub mod example;
+pub mod extra;
 pub mod gate;
 pub mod graph;
 pub mod meb;
